pyproject.toml
[project]
name = "totp-based-2fa-authentication-system-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "Pillow>=10.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pyotp>=2.9.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "qrcode>=8.2",
    "redis[asyncio]>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "starlette>=0.52.1",
    "uvicorn[standard]>=0.42.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers"
